Update pipeline input processor for db connection injection
Previously the datbase connection must be supplied via a mixin trait.
Although in principle this works, it conflicts with how we plan to supply
the database connection to our controllers. Database connection to the
controllers will be supplied via a constructor parameter, so it's easiest
if the pipeline processors also accept a constructor parameter.