Load resources as InputStream instead of File
Since they are not actual File objects in a filesystem (which File is
supposed to represent). Using them as File also causes problems when
we try to upload to newer versions of Jetty. We could create a fake
ZipFileSystem to keep using them as Files, but that's just not right.