Implement randomization for stats queries
The current implementation may be a problem when we have thousands
(or millions?) of items in the database to be returned. And indeed
there may be other ways to make this work from the database side:

- We could use mapReduce to assign random numbers and index on that

- We could assign random numbers when the records are created and
  sort on that.

However, none of the solutions are optimal:

- Using mapReduce is not only makes the queries less flexible (we have
  to evaluate JavaScript strings), but it could also be slower
  than the aggregation framework (our current implementation).

- Random numbers assignment does sort our items randomly, but only
  initially. Querying the endpoint again will not re-randomize
  the results.

And to top it off, randomization of query results is an issue
that the MongoDB team would want to solve
(https://jira.mongodb.org/browse/SERVER-533). So whatever we do
now, it's very likely that it will be made obsolete when database-level
randomization is implemented.

So for now, we stick with the simplest implementation: using Random.shuffle.